+++ Stalky106/Configs/StalkyPreferences.cs
@@ -1,13 +1,13 @@
-﻿using System;
-using System.ComponentModel;
-
-namespace Stalky106
+﻿namespace Stalky106.Configs
 {
-	[Serializable]
+	using System.ComponentModel;
+
 	public class StalkyPreferences
 	{
 		public bool AnnounceReady { get; set; } = true;
+
 		public float Cooldown { set; get; } = 40f;
+
 		public float InitialCooldown { set; get; } = 80f;
 
 		[Description("Should SCP-106 automatically teleport when stalking, or should he teleport manually?")]
@@ -19,19 +19,20 @@ namespace Stalky106
 		[Description("Forces SCP-106 to be teleported. If he's jumping, it will \"wait\" until he's on the ground to teleport him. Frame-perfect trick allows SCP-106 to move while teleporting, set this to false if players abuse it.")]
 		public bool ForceAutoTp { get; set; }
 
-		// Note: according to this benchmark: https://stackoverflow.com/a/10762995
-		// it's probably faster using an array instead of a set (maybe a tree set would help?)
 		[Description("ignore_teams and ignore_roles will ignore said teams and roles when searching for a player to stalk")]
-		public Team[] IgnoreTeams { set; get; } = new Team[] { Team.SCP, Team.CHI, Team.TUT };
-		public RoleType[] IgnoreRoles { set; get; } = new RoleType[] { RoleType.Scp106, RoleType.Scp079 };
+		public Team[] IgnoreTeams { set; get; } = { Team.SCP, Team.CHI, Team.TUT };
+
+		public RoleType[] IgnoreRoles { set; get; } =  { RoleType.Scp106, RoleType.Scp079 };
 		
 		[Description("Changes behaviour of stalk by only allowing to stalk players in the same zone as Larry")]
 		public bool SameZoneOnly { get; set; } = false;
 
 		[Description("The minimum amount of targetable players to be able to stalk.")]
 		public int MinimumAliveTargets { get; set; } = 0;
+
 		[Description("The minimum amount of alive players (any role) to be able to stalk.")]
 		public int MinimumAlivePlayers { get; set; } = 0;
+
 		[Description("The minimum amount of players connected in the server for Stalky to work.")]
 		public int MinimumPlayers { get; set; } = 0;
 	}
